вносиит правки по масштабу и слайдерам
