+++ js/main.js
@@ -3,6 +3,3 @@ import {renderPictures} from './picture.js';
 import './submit-form.js';
 
 renderPictures(similarPhotos);
-
-import './scale.js';
-import './newfilter.js';
