+++ js/slider.js
@@ -1,0 +1,50 @@
+const sliderElement = document.querySelector('.effect-level__slider');
+const imgPreview = document.querySelector('.img-upload__preview');
+//через стили это поле невидимое
+const effectValueInput = document.querySelector('.effect-level__value');
+const effectsList = document.querySelector('.effects__list');
+
+function getParamSlider(filter) {
+  return (
+    {
+      range: {
+        min: filter.min,
+        max: filter.max,
+      },
+      start: filter.start,
+      step: filter.step
+    });
+}
+
+function onUpdateSlider(filter) {
+  return function() {
+    effectValueInput.setAttribute('value', `${sliderElement.noUiSlider.get()}`);
+    imgPreview.style.filter = `${filter.filterEffect}(${effectValueInput.value}${filter.formatValue})`;
+  };
+}
+
+function desctroySlider() {
+  if (sliderElement.noUiSlider) {
+    sliderElement.noUiSlider.destroy();
+  }
+}
+
+function updateOptionsSlider(evt, filtersList) {
+  const currentFilter = filtersList[evt.target.value.toUpperCase()];
+  sliderElement.noUiSlider.updateOptions(getParamSlider(currentFilter));
+}
+
+function renderSlider(evt, currentNameFilter, filtersList) {
+  const currentFilter = filtersList[currentNameFilter];
+  if (!sliderElement.noUiSlider) {
+    noUiSlider.create(sliderElement, getParamSlider(currentFilter));
+  } else {
+    if (currentNameFilter !== 'NONE') {
+      effectsList.addEventListener('click', updateOptionsSlider(evt, filtersList));
+    }
+  }
+  // Обработчик событий на изменеие уровня слайдера
+  sliderElement.noUiSlider.on('update', onUpdateSlider(currentFilter));
+}
+
+export{getParamSlider, onUpdateSlider, desctroySlider, updateOptionsSlider, renderSlider};
